% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{scrape_chatgpt}
\alias{scrape_chatgpt}
\alias{scrape_chatgpt.default}
\alias{scrape_chatgpt.character}
\alias{scrape_chatgpt.data.frame}
\title{Scrape ChatGPT Conversations}
\usage{
scrape_chatgpt(x, ...)

\method{scrape_chatgpt}{default}(x, ...)

\method{scrape_chatgpt}{character}(x, ...)

\method{scrape_chatgpt}{data.frame}(x, column, ...)
}
\arguments{
\item{x}{A URL or data frame containing URLs to scrape.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A `data.frame` with the scraped conversation data.
}
\description{
A generic function to scrape ChatGPT conversations from a provided URL or data frame.
}
\examples{
scrape_chatgpt("https://chatgpt.com/share/b96f65e6-80c2-4c8b-940b-c9a98e048d95")

}
